requirements.txt
absl-py==1.4.0
addict==2.4.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
black==23.3.0
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
cityscapesScripts==2.2.2
click==8.1.3
cloudpickle==2.2.1
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.35
-e git+https://github.com/facebookresearch/detectron2.git@027099d37a246aebf835b0bd05b075758c3a40cd#egg=detectron2
docker-pycreds==0.4.0
einops==0.6.1
filelock==3.12.2
fonttools==4.39.4
fsspec==2023.6.0
ftfy==6.1.1
fvcore==0.1.5.post20221221
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.18.0
google-auth-oauthlib==1.0.0
grpcio==1.54.2
h5py==3.8.0
huggingface-hub==0.15.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
imageio==2.31.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
iopath==0.1.9
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mmcv==2.0.1
mmengine==0.8.2
MultiScaleDeformableAttention==1.0
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numpy==1.23.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
omegaconf==2.3.0
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.3
panopticapi @ git+https://github.com/cocodataset/panopticapi.git@7bb4655548f98f3fedc07bf37e9040a992b054b0
pathspec==0.11.1
pathtools==0.1.2
Pillow==9.5.0
platformdirs==3.5.1
portalocker==2.7.0
protobuf==3.19.0
psutil==5.9.5
ptflops==0.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
Pygments==2.15.1
pyparsing==3.0.9
pyquaternion==0.9.9
python-dateutil==2.8.2
pytorch-model-summary==0.1.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.6.3
requests==2.30.0
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scipy==1.10.1
sentry-sdk==1.28.1
setproctitle==1.3.2
shapely==2.0.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.0
submitit==1.4.5
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.0
termcolor==2.3.0
thop==0.1.1.post2209072238
tifffile==2023.4.12
timm==0.5.4
tomli==2.0.1
torch==1.10.1+cu111
torchaudio==0.10.1
torchstat==0.0.7
torchvision==0.11.2+cu111
tqdm==4.65.0
typing==3.7.4.3
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1685704949284/work
tzdata==2023.3
urllib3==1.26.15
wandb==0.15.5
wcwidth==0.2.6
Werkzeug==2.3.4
yacs==0.1.8
yapf==0.40.1
zipp==3.15.0
